COM-1296 add profile picture to accountInfo
