+++ src/core/pages/AccountInfo.vue
@@ -1,7 +1,23 @@
 <template>
   <q-page :class="backgroundClass" padding>
-    <h4>Mon compte</h4>
+    <div>
+      <h4>Mon compte</h4>
+      <div class="q-mb-xl">
+        <div class="photo-caption">Photo</div>
+        <div class="row gutter-profile">
+          <div class="col-xs-12 col-md-6">
+            <ni-picture-uploader :user-profile="mergedUserProfile" :background="backgroundClass"
+              :refresh-picture="refreshUser" :key="pictureRenderTimes" />
+          </div>
+        </div>
+      </div>
       <div class="row gutter-profile q-mb-xl">
+        <ni-input caption="Prénom" :error="$v.mergedUserProfile.identity.firstname.$error"
+          v-model.trim="mergedUserProfile.identity.firstname" @blur="updateUser('identity.firstname')"
+          @focus="saveTmp('identity.firstname')" />
+        <ni-input caption="Nom" :error="$v.mergedUserProfile.identity.lastname.$error"
+          v-model.trim="mergedUserProfile.identity.lastname" @blur="updateUser('identity.lastname')"
+          @focus="saveTmp('identity.lastname')" />
         <div class="col-xs-12 col-md-6 row items-center">
           <div class="col-11">
             <ni-input ref="userEmail" name="emailInput" caption="Email" type="email" :disable="emailLock"
@@ -30,6 +46,7 @@
         </div>
         <div class="cursor-pointer"><a @click.prevent="rgpdModal = true">Politique RGPD</a></div>
       </div>
+    </div>
 
     <!-- New password modal -->
     <ni-modal v-model="newPasswordModal" @hide="resetForm">
@@ -57,6 +74,7 @@
 </template>
 
 <script>
+import 'vue-croppa/dist/vue-croppa.css'
 import { required, requiredIf, email, sameAs } from 'vuelidate/lib/validators';
 import get from 'lodash/get';
 import set from 'lodash/set';
@@ -66,6 +84,7 @@ import Input from '@components/form/Input';
 import HtmlModal from '@components/modal/HtmlModal';
 import Modal from '@components/modal/Modal';
 import { NotifyWarning, NotifyPositive, NotifyNegative } from '@components/popup/notify';
+import PictureUploader from '@components/PictureUploader.vue';
 import { frPhoneNumber } from '@helpers/vuelidateCustomVal';
 import { passwordMixin } from '@mixins/passwordMixin';
 import { validationMixin } from 'src/modules/client/mixins/validationMixin';
@@ -81,12 +100,18 @@ export default {
     'ni-input': Input,
     'ni-html-modal': HtmlModal,
     'ni-modal': Modal,
+    'ni-picture-uploader': PictureUploader,
   },
   data () {
     return {
       mergedUserProfile: {
+        identity: {
+          firstname: '',
+          lastname: '',
+        },
         local: { email: '', password: '' },
         contact: { phone: '' },
+        picture: { link: '' },
       },
       tmpInput: '',
       emailLock: true,
@@ -99,11 +124,16 @@ export default {
       cguCompani,
       backgroundClass: /\/ad\//.test(this.$router.currentRoute.path) ? 'vendor-background' : 'client-background',
       isLoggingOut: false,
+      pictureRenderTimes: 0,
     }
   },
   validations () {
     return {
       mergedUserProfile: {
+        identity: {
+          firstname: { required },
+          lastname: { required },
+        },
         local: {
           email: { required, email },
           password: { required, ...this.passwordValidation },
@@ -119,17 +149,28 @@ export default {
     }
   },
   async created () {
-    try {
-      const user = await Users.getById(this.$route.params.id);
-      this.mergedUserProfile = { contact: {}, ...pick(user, ['local.email', 'contact']) };
-    } catch (e) {
-      console.error(e);
-    }
+    await this.refreshUser();
   },
   async beforeDestroy () {
     if (this.isLoggingOut) this.$store.dispatch('main/resetMain');
   },
   methods: {
+    async refreshUser () {
+      try {
+        const user = await Users.getById(this.$route.params.id);
+        this.mergedUserProfile = {
+          contact: {},
+          ...pick(user,
+            ['_id', 'identity.firstname', 'identity.lastname', 'picture.link', 'local.email', 'contact', 'role']),
+        };
+        if (!get(this.mergedUserProfile, 'picture.link')) this.forcePictureNewRender();
+      } catch (e) {
+        console.error(e);
+      }
+    },
+    forcePictureNewRender () {
+      this.pictureRenderTimes += 1;
+    },
     saveTmp (path) {
       if (this.tmpInput === '') this.tmpInput = get(this.mergedUserProfile, path);
     },
@@ -198,4 +239,7 @@ export default {
   @media screen and (max-width: 676px)
     display: flex
     justify-content: center
+.photo-caption
+  font-size: 12px
+  margin: 0 0 4px 0
 </style>
